move commit logic to cli
